Ensure the coroutines debug agent attaches even with no stdlib
With this change, we ensure that `AgentPremain` does not touch
any of the `kotlin.*` APIs when it's being attached to a process.
This means that even if it's attached by a mistake to a pure Java
process, it won't cause any harm.

Most of the changes are limited to removing non-idiomatic
`runCatching` calls, but one sad consequence is the inability
to use the convenient `thread { }` function.

An integration test was implemented checking that a
pure Java process completes successfully even with the debug agent
attached to it.
The earlier version of the test performed the check even before
the Java class got compiled, which was fine, given that
the execution never even got to the main class and crashed during
the debug agent attach, so it had to be significantly changed.