Move the test sources from the JVM to the concurrent source set
See the commit directly following this one in history for the
context behind the change.
The reason to split this commit off is so that `git` recognizes
the following commit as changes to existing files instead of
the files getting removed and created separately.
This helps preserve the change history.